requirements.txt
requests
psycopg2
python-dotenv
beautifulsoup4
neo4j
Flask
gunicorn
pandas